refactor: move env base modules to unilab/base
Restructure codebase by moving environment base classes and utilities from unilab/envs to unilab/base for better separation of concerns. This includes backend implementations, base classes, curriculum learning, dtype config, np_env, and registry modules.
