+++ scripts/train_mlx_ppo.py
@@ -46,7 +46,7 @@ def ensure_registries() -> None:
 ensure_registries()
 
 from unilab.config import locomotion_params
-from unilab.envs import registry
+from unilab.base import registry
 from unilab.utils import render_many
 from unilab.utils.onpolicy_logger import OnPolicyLogger
 from unilab.algos.mlx.common import EmpiricalDiscountedVariationNormalization, RolloutBuffer
@@ -159,7 +159,7 @@ def play_mlx_ppo(args, cfg, dtype, use_fp16, resolved_sim_backend, task_log_root
     """Play mode for MLX PPO."""
     import mlx.core as mx
     import numpy as np
-    from unilab.envs import registry
+    from unilab.base import registry
     from unilab.utils import render_many
 
     play_model_dtype = mx.float32 if use_fp16 else dtype
