+++ unilab/base/__init__.py
@@ -1,0 +1,1 @@
+"""Environment registry and base classes."""
