+++ unilab/envs/locomotion/go2/joystick.py
@@ -5,9 +5,9 @@ import gymnasium as gym
 import numpy as np
 from dataclasses import dataclass, field
 
-from unilab.envs import registry
-from unilab.envs.np_env import NpEnvState
-from unilab.envs.backend import create_backend
+from unilab.base import registry
+from unilab.base.np_env import NpEnvState
+from unilab.base.backend import create_backend
 from unilab.utils.math_utils import np_quat_mul, np_yaw_to_quat
 
 from unilab.envs.locomotion.go2.base import Go2BaseEnv, Go2BaseCfg
